refactor: reorganize bundle code
